Moved ProductSelect to its own test class
